docs: fix styles, typos, ordering
