+++ site/snippets/getting-started/setup-app/route.ts
@@ -1,10 +1,9 @@
+import { arbitrumSepolia } from "@alchemy/aa-core";
 import { NextResponse } from "next/server";
 
-export async function POST(
-  req: Request,
-  { params }: { params: { routes: string[] } }
-) {
-  const apiUrl = "https://api.g.alchemy.com";
+export async function POST(req: Request) {
+  // if you want to handle other or multiple chains, you can change this line
+  const rpcUrl = arbitrumSepolia.rpcUrls.alchemy.http[0];
   const apiKey = process.env.ALCHEMY_API_KEY;
 
   if (apiKey == null) {
@@ -16,10 +15,9 @@ export async function POST(
 
   const body = await req.json();
 
-  const res = await fetch(apiUrl + `/${params.routes.join("/")}`, {
+  const res = await fetch(`${rpcUrl}/${apiKey}`, {
     method: "POST",
     headers: {
-      Authorization: `Bearer ${apiKey}`,
       ...req.headers,
     },
     body: JSON.stringify(body),
